+++ ts/src/lib/video/video-codec.ts
@@ -2,11 +2,14 @@ import { CaseStyle } from "kryo/case-style";
 import { TsEnumType } from "kryo/types/ts-enum";
 
 export enum VideoCodec {
-  SorensonH263,
+  None,
+  Jpeg,
+  Sorenson,
   Screen,
   Vp6,
   Vp6Alpha,
   Screen2,
+  Avc,
 }
 
 export const $VideoCodec: TsEnumType<VideoCodec> = new TsEnumType<VideoCodec>({